Add ClientRepository and save logic to ReservationService.java
ClientRepository was added to ReservationService.java to get access to the client data. Additionally, save logic was added to the 'save' method within ReservationService. This change allows the system to store a reservation into the database with the appropriate associated client and car entities, improving data reliability and consistency. The ClientRepository file was also created as a separate interface to fetch client information as needed.