Improve error handling and code readability in ReservationService
This commit modifies the method save() in ReservationService.java. It changes the way we fetch ClientEntity and CarEntity by using orElseThrow(), which will help to handle the NoSuchElementException at Runtime. It improves error handling when there is no client or car found with the given details. Furthermore, in the save method, the carRepository.save() was redundant and replaced with reservationRepository.save() to correctly save the reservation. Lastly, it updates the reservation object's id with the saved entity's id.

Additionally, import statements and save() method in PaymentService.java are adjusted to work with these changes, adding necessary functionality to store payments in the database.