Refactor PaymentRepository's findByUsername method
The import of ReservationEntity in PaymentRepository was removed as it was unnecessary. The method findByClientUsername in PaymentRepository is adjusted to findByReservation_Client_Username to align with changes in the ReservationService. This improves the ability to fetch payments by the username of the reservation client and not just the client, thus adding precision to the query and making the code base more efficient.

These changes are mirrored in PaymentService.java where how we call the method is adjusted accordingly. This provides more accurate results when searching for payments by a user.