Add Docker support for frontend and backend
This commit introduces Docker support to the project for both the frontend and backend. We create `.dockerignore` files for each, specifying the files and directories to be ignored during the build of the Docker images. We also add a `docker-compose.yml` file to manage the multi-container application. This allows the frontend and backend to be built and run in separate containers, making the application easier to deploy and scale.