Refactor AuthbootClient for dynamic Url setting
Changed the hardcoded localhost URL in AuthbootClient to be dynamically set based on the application configuration.
This allows the Feign client used for communication with the auth service to fetch the Url directly from `spring.properties.feign.auth`.
This is important for increasing flexibility especially when deploying in different environments.
Additional changes were made in docker-compose.yml and application.yml to support this feature.